+++ basicalter/number_test.go
@@ -1,9 +1,15 @@
 package basicalter_test
 
 import (
+	"math/bits"
 	"testing"
 
 	"github.com/jeremmfr/go-utils/basicalter"
+)
+
+const (
+	maxInt = 1<<(bits.UintSize-1) - 1
+	minInt = -1 << (bits.UintSize - 1)
 )
 
 func TestAbsolutInt(t *testing.T) {
@@ -15,4 +21,15 @@ func TestAbsolutInt(t *testing.T) {
 	if v := basicalter.AbsoluteInt(basicalter.AbsoluteInt(i)); v < 0 {
 		t.Errorf("AbsoluteInt(AbsoluteInt(%d)) return negative integer: %d", i, v)
 	}
+
+	i = minInt + 1
+	if v := basicalter.AbsoluteInt(i); v != maxInt {
+		t.Errorf("AbsoluteInt(%d) not return expected integer %d, found %d", i, maxInt, v)
+	}
+}
+
+func TestAbsolutIntPanic(t *testing.T) {
+	defer func() { _ = recover() }()
+	basicalter.AbsoluteInt(minInt)
+	t.Errorf("AbsoluteInt doesn't panic with the minimum value of integers")
 }